Title: 실수 입력받아 계산, Time: 6ms, Memory: 0MB, Status: Passed - Codetree

+++ 250505/실수 입력받아 계산/calculation-by-inputting-a-real-number.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int a, b, sum;
+    double a, b, sum;
 
     cin>>a>>b;
